requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
google-generativeai>=0.8.5
google-genai>=1.47.0
protobuf>=3.20.2,<6.0.0
qdrant-client==1.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
pydantic>=2.12.5
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
httpx>=0.28.1
